requirements.txt
roslibpy
python-geometry-msgs
sensor_msgs.msg 
readchar